proper_case v1.1.0 ProperCase

An Elixir library that converts keys in maps between snake_case and camel_case

Functions

camel_case(key)

Converts a string to camelCase

snake_case(val)

Converts a string to snake_case

to_camel_case(map)

Converts all the keys in a map to camelCase. If the map is a struct with no Enumerable implementation, the struct is considered to be a single value.

to_snake_case(map)

Converts all the keys in a map to snake_case. If the map is a struct with no Enumerable implementation, the struct is considered to be a single value.